calib: Added entire param list to log
In the past, the log kept track only of the list of parameters that were
being calibrated over. This was annoying if you wanted to shove those
parameters into lake.m.

Now the log file has two lines: the first is just the varied parameters,
the second if the entire set of numbers fed to lake.m in the objective
function calculation